django_backend/requirements.txt
django-cors-headers==3.4.0
djangorestframework==3.11.0
Django==3.0.7
djangorestframework-simplejwt==4.4.0
